move xdr specific functions to xdr_decoder.go
